Add company and review features to the application
This commit introduces company and review features to the application. Specifically, it adds the ability to fetch a list of all companies and reviews. In addition to that, it offers to rename Companies files. This change also includes components for a company card and the creation of new tab components.